requirements.txt
fastapi
pydantic
sympy
numpy
python-multipart
python-dotenv
requests
uvicorn
jinja2
markupsafe
pytesseract
Pillow
google-generativeai
